Download the data set
We will use an insurance data set from Kaggle. You can find it here. Click on the Download button, and you should see that you will download a file named health-insurance-cross-sell-prediction.zip. Once you unzip the file, you should see train.csv. This is the data set we will use for the remainder of the example. Remember that this example is purely educational, and you could use any data set you want - we just happened to choose this one.